move TorchWriter back to tensorboard
