+++ harness/determined/tensorboard/base.py
@@ -132,9 +132,9 @@ def get_metric_writer() -> tensorboard.BatchMetricWriter:
 
     except ModuleNotFoundError:
         logging.warning("TensorFlow writer not found")
-        from determined.pytorch.tensorboard import TorchWriter
+        from determined.tensorboard.metric_writers.pytorch import _TorchWriter
 
-        writer = TorchWriter()
+        writer = _TorchWriter()
 
     return tensorboard.BatchMetricWriter(writer)
